# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Kavling:
    properties:
      created_at:
        type: string
      id:
        type: string
      nama:
        type: string
      updated_at:
        type: string
    type: object
  model.KavlingInput:
    properties:
      nama:
        type: string
    required:
    - nama
    type: object
  model.Perlengkapan:
    properties:
      created_at:
        type: string
      deskripsi:
        type: string
      harga:
        type: integer
      id:
        type: string
      nama:
        type: string
      stok:
        type: integer
      updated_at:
        type: string
    type: object
  model.PerlengkapanInput:
    properties:
      deskripsi:
        type: string
      harga:
        type: integer
      nama:
        type: string
      stok:
        type: integer
    required:
    - deskripsi
    - harga
    - nama
    - stok
    type: object
...
  utils.ErrorResponseData:
    properties:
      data:
        description: 'Example: nil'
      message:
        description: |-
          The error message.
          Example: "Invalid request"
        type: string
    type: object
  utils.SuccessResponseData:
    properties:
      data:
        description: The actual data returned.
      message:
        description: |-
          The success message.
          Example: "success"
        type: string
    type: object
host: api.perkemahanbedengan.com
info:
  contact:
    email: adityarizky1020@gmail.com
    name: Aditya Rizky Ramadhan
  description: |-
    This is API documentation for Islamind. You can use the API with the following hosts:
    - Production: `api.perkemahanbedengan.com`
  title: Islamind API
  version: "1.0"
paths:
  /kavling:
    get:
      consumes:
      - application/json
      description: Menampilkan semua data kavling
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Kavling'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menampilkan semua data kavling
      tags:
      - Kavling
    post:
      consumes:
      - application/json
      description: Membuat data kavling baru
      parameters:
      - description: Data Kavling
        in: body
        name: kavling
        required: true
        schema:
          $ref: '#/definitions/model.KavlingInput'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Kavling'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Membuat data kavling baru
      tags:
      - Kavling
  /kavling/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data kavling berdasarkan id
      parameters:
      - description: ID Kavling
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Kavling
        in: body
        name: kavling
        required: true
        schema:
          $ref: '#/definitions/model.KavlingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Kavling'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menghapus data kavling berdasarkan id
      tags:
      - Kavling
    get:
      consumes:
      - application/json
      description: Menampilkan data kavling berdasarkan id
      parameters:
      - description: ID Kavling
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Kavling'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menampilkan data kavling berdasarkan id
      tags:
      - Kavling
    put:
      consumes:
      - application/json
      description: Memperbarui data kavling berdasarkan id
      parameters:
      - description: ID Kavling
        in: path
        name: id
        required: true
        type: string
      - description: Data Kavling
        in: body
        name: kavling
        required: true
        schema:
          $ref: '#/definitions/model.KavlingInput'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Kavling'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Memperbarui data kavling berdasarkan id
      tags:
      - Kavling
  /perlengkapan:
    get:
      consumes:
      - application/json
      description: Menampilkan semua data perlengkapan
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Perlengkapan'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menampilkan semua data perlengkapan
      tags:
      - Perlengkapan
    post:
      consumes:
      - application/json
      description: Membuat data perlengkapan baru
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Perlengkapan data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PerlengkapanInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Perlengkapan'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Membuat data perlengkapan baru
      tags:
      - Perlengkapan
  /perlengkapan/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data perlengkapan berdasarkan id
      parameters:
      - description: ID Perlengkapan
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menghapus data perlengkapan berdasarkan id
      tags:
      - Perlengkapan
    get:
      consumes:
      - application/json
      description: Menampilkan data perlengkapan berdasarkan id
      parameters:
      - description: ID Perlengkapan
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Perlengkapan'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menampilkan data perlengkapan berdasarkan id
      tags:
      - Perlengkapan
    put:
      consumes:
      - application/json
      description: Memperbarui data perlengkapan berdasarkan id
      parameters:
      - description: ID Perlengkapan
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Perlengkapan data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PerlengkapanInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Perlengkapan'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Memperbarui data perlengkapan berdasarkan id
      tags:
      - Perlengkapan
...
swagger: "2.0"